+++ server/app/Http/Controllers/ItemVoteController.php
@@ -7,16 +7,17 @@ use App\Models\BoardItem;
 use App\Models\ItemVote;
 use App\Models\Workspace;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class ItemVoteController extends Controller
 {
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Workspace $workspace, Board $board, BoardItem $item)
+    public function store(Workspace $workspace, Board $board, BoardItem $item, Request $request)
     {
         $itemVote = ItemVote::createOrFirst([
-            'user_id' => auth()->user()->id,
+            'user_id' => Auth::user()->id,
             'board_item_id' => $item->id
         ]);
 
@@ -26,8 +27,12 @@ class ItemVoteController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Workspace $workspace, Board $board, BoardItem $item, ItemVote $vote)
+    public function destroy(Workspace $workspace, Board $board, BoardItem $item, ItemVote $vote, Request $request)
     {
+        if ($request->user()->cannot('delete', $vote)) {
+            abort(403);
+        }
+
         $vote->delete();
 
         return response()->noContent();
