+++ server/app/Http/Controllers/WorkspaceController.php
@@ -4,35 +4,39 @@ namespace App\Http\Controllers;
 
 use App\Models\Workspace;
 use App\Http\Requests\StoreWorkspaceRequest;
-use App\Http\Requests\UpdateWorkspaceRequest;
 use App\Models\User;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Nette\NotImplementedException;
-use Symfony\Component\Translation\Exception\NotFoundResourceException;
 
 class WorkspaceController extends Controller
 {
-    public function index() {
+    public function index()
+    {
         $authUser = Auth::user();
         $workspaces = Workspace::where('user_id', $authUser->id)->orderBy('created_at', 'asc')->get();
 
         return response()->json(["data" => $workspaces]);
     }
 
-    public function show(Workspace $workspace) {
+    public function show(Workspace $workspace)
+    {
         return response()->json(["data" => $workspace]);
     }
 
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreWorkspaceRequest $request)
+    public function store(Request $request)
     {
-        $authUser = Auth::user();
-        $user = User::find($authUser->id);
-        $params = $request->validated();
-        $params['user_id'] = $user->id;
-        $workspace = $user->workspaces()->create($params);
+        $validatedData = $request->validate([
+            'name' => ['required'],
+            'subdomain' => ['required', 'min:1', 'max:255']
+        ]);
+
+        $user = Auth::user();
+
+        $validatedData['user_id'] = $user->id;
+        $workspace = $user->workspaces()->create($validatedData);
 
         return response()->json(["data" => $workspace]);
     }
@@ -40,10 +44,16 @@ class WorkspaceController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(int $workspaceId, UpdateWorkspaceRequest $updateWorkspaceRequest)
+    public function update(Workspace $workspace, Request $request)
     {
-        $params = $updateWorkspaceRequest->validated();
-        $workspace = Workspace::findOrFail($workspaceId);
+        if ($request->user()->cannot('update', $workspace)) {
+            abort(403);
+        }
+
+        $params = $request->validate([
+            'name' => ['required'],
+            'subdomain' => ['required', 'min:1', 'max:255'],
+        ]);
 
         $workspace->name = $params["name"];
         $workspace->subdomain = $params["subdomain"];
@@ -56,12 +66,14 @@ class WorkspaceController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(int $workspaceId)
+    public function destroy(Workspace $workspace, Request $request)
     {
-        $workspace = Workspace::findOrFail($workspaceId);
+        if ($request->user()->cannot('delete', $workspace)) {
+            abort(403);
+        }
 
         $workspace->delete();
 
-        return response()->json(["data" => ["message" => $workspace->name . " has been deleted successfully."]]);
+        return response()->json(["data" => ["message" => "{$workspace->name} has been deleted successfully."]]);
     }
 }
